Add init(ResizeHandler) function to SceneryPanel for initialisation of resizing handlers
